pyproject.toml
[tool.poetry]
name = "tiled-heightmap-renderer"
version = "0.1.0"
description = ""
authors = ["Elliot <3186037+elliot-100@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.5.0"

[tool.poetry.group.test.dependencies]
mypy = "^1.10.0"
pytest = "^8.2.2"

[tool.mypy]
strict = true
warn_unused_configs = true
enable_error_code = ["ignore-without-code"]  # Require specific codes for ignores
disallow_any_generics = false

[tool.ruff]
target-version = "py310"  # Ruff doesn't respect Python dependency in [tool.poetry.dependencies]

[tool.ruff.lint]
select = ["ALL"]

ignore = [
# Ignore rules that conflict with Ruff formatter:
#   COM812 Trailing comma missing
#   ISC001 Implicitly concatenated string literals on one line
"COM812", "ISC001",

# Ignore other rules:
#   ANN101 Missing type annotation for `self` in method
#   ANN102 Missing type annotation for `cls` in classmethod
#   D205 1 blank line required between summary line and description
#   PLR0913 Too many arguments in function definition
 "ANN101", "ANN102", "D205", "PLR0913",
]

[tool.ruff.lint.per-file-ignores]
# Ignore rules that aren't relevant in tests:
#   S101 Use of assert detected
#   PLR2004 Magic value used in comparison
"**/{tests}/*" = ["S101", "PLR2004"]
# Don't check line-lengths in example heightmaps
"heightmap_renderer/__init__.py" = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
